Extract TripSummaryUtil and refactor SLAP
